% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_tab.R
\name{anova_tab}
\alias{anova_tab}
\title{Format an analysis of variance table}
\usage{
anova_tab(mod, type = 3, pretty_names = NULL)
}
\arguments{
\item{mod}{A linear model object created with lm().}

\item{type}{A number 2 or 3 to indicate the type of sums of squares to use in the analysis. Defaults to 3.}

\item{pretty_names}{An optional vector to replace column names with formatted variable names.}
}
\value{
Formatted ANOVA table
}
\description{
Format an analysis of variance table
}
