% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_tab.R
\name{corr_tab}
\alias{corr_tab}
\title{Format a bivariate correlation table}
\usage{
corr_tab(x, pretty_names = NULL, num_cols = F)
}
\arguments{
\item{x}{A data frame, matrix, or tibble containing only numeric columns to be included in the bivariate correlations.}

\item{pretty_names}{An optional vector with formatted dependent variable names.}

\item{num_cols}{Logical, when TRUE numbers each variable and replaces column names with corresponding numbers. Defaults to FALSE.}
}
\value{
Formatted bivariate correlation table
}
\description{
Format a bivariate correlation table
}
