% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etaSquared.R
\name{etaSquared}
\alias{etaSquared}
\title{Find eta squared and partial eta squared for analysis of variance.}
\usage{
etaSquared(x, type = 3, anova = FALSE)
}
\arguments{
\item{x}{A linear model object created with lm().}

\item{type}{A number 2 or 3 to indicate the type of sums of squares to use in the analysis.}

\item{anova}{Logical, indicating whether or not to include full ANVOA output with effect size.}
}
\value{
Eta squared and partial eta squared, ANOVA table if requested
}
\description{
Find eta squared and partial eta squared for analysis of variance.
}
