% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hreg_tab.R
\name{hreg_tab}
\alias{hreg_tab}
\title{Format hierarchical linear regression table}
\usage{
hreg_tab(mod_steps, preds_vec, betas = "unstd", se = T)
}
\arguments{
\item{mod_steps}{A list of linear model objects created with lm(), in the order of analysis.}

\item{preds_vec}{A vector of strings indicating variable names. Strings do not need to match column names, but do need to be in the same order as the independent variables in the call for mod.}

\item{betas}{"unstd" or "std" Indicates if you want unstandardized coefficients or standardized beta coefficients. Defaults to c("unstd", "std").}

\item{se}{Logical, indicates if you want standard errors. Defaults to TRUE.}
}
\value{
Formatted hierarchical regression table
}
\description{
Format hierarchical linear regression table
}
