% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_tab.R
\name{reg_tab}
\alias{reg_tab}
\title{Format linear regression table}
\usage{
reg_tab(mod, preds_vec, betas = c("unstd", "std"), se = T)
}
\arguments{
\item{mod}{A linear model object created with lm().}

\item{preds_vec}{A vector of strings indicating variable names. Strings do not need to match column names, but do need to be in the same order as the independent variables in the call for mod.}

\item{betas}{"unstd" or "std" Indicates if you want unstandardized coefficients or standardized beta coefficients. Defaults to c("unstd", "std").}

\item{se}{Logical, indicates if you want standard errors. Defaults to TRUE.}
}
\value{
Formatted regression table
}
\description{
Format linear regression table
}
