% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signif_stars.R
\name{signif_stars}
\alias{signif_stars}
\title{Add significance stars to an estimate}
\usage{
signif_stars(estimate, p_value)
}
\arguments{
\item{estimate}{A point estimate or statistic, should be a single numeric value.}

\item{p_value}{A p-value from a statistical test, should be a single numeric value between 0 and 1.}
}
\value{
The estimate with the appropriate number of stars, depending on level of significance.
}
\description{
Add significance stars to an estimate
}
